// src/lib/api/models/getApiLeadsSource.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiLeadsSource = typeof GetApiLeadsSource[keyof typeof GetApiLeadsSource];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiLeadsSource = {
  WEBSITE: 'WEBSITE',
  PROPERTY_DETAIL: 'PROPERTY_DETAIL',
  OWNER_PAGE: 'OWNER_PAGE',
  SALES_LINK: 'SALES_LINK',
  CTV_LINK: 'CTV_LINK',
  AGENCY_MARKETING: 'AGENCY_MARKETING',
  MANUAL_INPUT: 'MANUAL_INPUT',
  LEAD_POOL: 'LEAD_POOL',
  IMPORT: 'IMPORT',
} as const;